# opt/vpp-restapi/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /vpp/interfaces:
    get:
      description: Get all VPP interfaces with status.
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all interfaces
      tags:
      - interfaces
...
swagger: "2.0"